server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy
psycopg2-binary
pyyaml
Flask-Migrate
gunicorn
flask-restful
flask-login